+++ domain/src/main/java/com/myapp/chatmemo/domain/model/value/RoomId.kt
@@ -1,3 +1,3 @@
 package com.myapp.chatmemo.domain.model.value
 
-inline class RoomId(val value: Int)
+inline class RoomId(val value: Long)
