+++ domain/src/main/java/com/myapp/chatmemo/domain/model/value/TemplateId.kt
@@ -1,3 +1,3 @@
 package com.myapp.chatmemo.domain.model.value
 
-inline class TemplateId(val value: Int)
+inline class TemplateId(val value: Long)
